Add files via upload
These are both of the models I created for the MPG model. I made a slight typo the other night because I input the data incorrectly. You'll get a negative mpg if you put in the stats for a car created in 2000 or newer because years are handled in xx format instead of xxxx. Too lazy to fix.